Shift Summary class to its own file
