+++ src/main/java/loanbook/logic/commands/SummaryCommand.java
@@ -8,7 +8,6 @@ import java.util.List;
 import loanbook.logic.CommandHistory;
 import loanbook.model.Model;
 import loanbook.model.loan.Loan;
-import loanbook.model.loan.LoanStatus;
 
 /**
  * Terminates the program.
@@ -33,68 +32,5 @@ public class SummaryCommand extends Command {
         }
 
         return new CommandResult(String.format(MESSAGE_SUMMARY_ACKNOWLEDGEMENT, summary.getSummary()));
-    }
-}
-
-/**
- * Class to encapsulate all the statistics to be kept track of.
- */
-class Summary {
-    public static final String MESSAGE_SUMMARY = "You have loaned %1$d loan(s). "
-            + "You have %2$d loan(s) ongoing.\n"
-            + "Your total revenue is $%3$.2f.";
-
-    private int numLoans;
-    private int numLoansInProgress;
-    private double totalRevenue;
-
-    public Summary() {
-        numLoans = 0;
-        numLoansInProgress = 0;
-        totalRevenue = 0;
-    }
-
-    /**
-     * Adds the statistics of a loan into the summary object
-     * @param loan
-     */
-    public void addLoan(Loan loan) {
-        addNumLoans();
-        if (loan.getLoanStatus() == LoanStatus.ONGOING) {
-            addNumLoansInProgress();
-        } else {
-            addTotalRevenue(Math.floor(loan.calculateCost() * 100) / 100);
-        }
-    }
-
-    public int getNumLoans() {
-        return numLoans;
-    }
-
-    public void addNumLoans() {
-        numLoans++;
-    }
-
-    public int getNumLoansInProgress() {
-        return numLoansInProgress;
-    }
-
-    public void addNumLoansInProgress() {
-        numLoansInProgress++;
-    }
-
-    public double getTotalRevenue() {
-        return totalRevenue;
-    }
-
-    public void addTotalRevenue(double revenue) {
-        totalRevenue += revenue;
-    }
-
-    public String getSummary() {
-        return String.format(MESSAGE_SUMMARY,
-                getNumLoans(),
-                getNumLoansInProgress(),
-                getTotalRevenue());
     }
 }
